tests: Move view tests to their own directory
This lets our unit tests mirror the main code structure and highlights
gaps in coverage.
